add registration property to SignUp response
